Show the error message if the OAuth failed
If the Connect to QuickBooks flow failed, this shows the error message received on the page so that the user can be alerted to fix the problem. Before this change, the callback page ignored the error and attempted to get a token anyway, resulting in a confusing PHP error message (see #2).